requirements.txt
adbutils==0.11.0
allure-pytest==2.8.6
allure-python-commons==2.8.6
allure-robotframework==2.8.6
apipkg==1.5
apkutils2==1.0.0
Appium-Python-Client==1.1.0
asn1crypto==1.0.1
atomicwrites==1.3.0
attrs==19.3.0
bcrypt==3.1.7
cached-property==1.5.2
certifi==2019.9.11
cffi==1.12.3
chardet==3.0.4
cigam==0.0.3
colorama==0.4.3
cryptography==2.7
cycler==0.10.0
Cython==0.29.22
decorator==4.4.2
Deprecated==1.2.12
deprecation==2.1.0
enum34==1.1.6
et-xmlfile==1.0.1
execnet==1.7.1
facebook-wda==1.4.1
html5lib==1.0.1
idna==2.7
imageio==2.9.0
images2gif==1.0.1
importlib-metadata==0.23
ipython-genutils==0.2.0
jdcal==1.4.1
Jinja2==2.10
kiwisolver==1.3.1
logzero==1.7.0
lxml==4.4.2
MarkupSafe==1.1.1
matplotlib==3.3.4
mock==4.0.3
more-itertools==7.2.0
multidict==4.7.5
namedlist==1.7
nose==1.3.7
numpy==1.19.5
opencv-python==4.5.1.48
openpyxl==3.0.0
packaging==19.2
paramiko==2.6.0
parso==0.4.0
pexpect==4.7.0
pickleshare==0.7.5
Pillow==8.1.1
pithy==0.0.3
pluggy==0.11.0
proglog==0.1.9
progress==1.5
prompt-toolkit==2.0.9
ptyprocess==0.6.0
py==1.8.0
pycparser==2.19
pyelftools==0.27
Pygments==2.4.2
pymongo==3.6.1
PyMySQL==0.9.3
PyNaCl==1.3.0
pynput==1.7.3
pyparsing==2.4.5
pypiwin32==223
pytesseract==0.2.4
pytest==4.5.0
pytest-forked==1.1.3
pytest-html==1.20.0
pytest-metadata==1.8.0
pytest-ordering==0.6
pytest-rerunfailures==7.0
pytest-xdist==1.31.0
python-dateutil==2.8.0
pywin32==228
PyYAML==5.1.1
redis==2.10.6
requests==2.22.0
retry==0.9.2
robotframework==3.1.2
robotframework-seleniumlibrary==4.0.0
selenium==3.141.0
six==1.12.0
text-unidecode==1.2
tornado==6.1
tqdm==4.32.1
traitlets==4.3.2
uiautomator2==2.16.0
urllib3==1.25.2
violent-webdriver==1.0.28
wcwidth==0.1.7
webencodings==0.5.1
weditor==0.6.4
whichcraft==0.6.1
wrapt==1.12.1
xlrd==1.2.0
xlutils==2.0.0
xlwt==1.3.0
xmltodict==0.12.0
yarl==1.3.0
yml==0.0.1
zipp==0.6.0
